blog_backend/requirements.txt
Django==5.2.4
PyJWT==2.8.0
django-cors-headers==4.3.1
